requirements.txt
git+https://github.com/Heat-Miser/xlrd@master
yara-python
